Changes license from CC BY-SA to GNU GPLv3 or later
Using CC BY-SA for software is not recommended even by the Creative Commons
organization [0]. The GNU GPLv3 or later has at least the same requirements
as the CC BY-SA license, but is more suited for software in that it specifically
mentions source code and object code.

The use of the LGPL over the GPL is negligible in this case as there is no
concept of dynamic linking for microcontroller firmware.

[0] http://wiki.creativecommons.org/Frequently_Asked_Questions#Can_I_apply_a_Creative_Commons_license_to_software.3F

Signed-off-by: ZachWick <zach@zachwick.com>